+++ ethereum/core/src/test/java/org/hyperledger/besu/ethereum/core/TransactionSizeAndHashTest.java
@@ -307,6 +307,11 @@ public class TransactionSizeAndHashTest {
   }
 
   private static Transaction buildBlobTransaction() {
+    final KZGCommitment commitment = new KZGCommitment(Bytes48.fromHexStringLenient("0x0987"));
+    final VersionedHash versionedHash =
+        new VersionedHash(
+            VersionedHash.SHA256_VERSION_ID,
+            Hash.wrap(org.hyperledger.besu.crypto.Hash.sha256(commitment.getData())));
     return Transaction.builder()
         .type(TransactionType.BLOB)
         .chainId(BigInteger.ONE)
@@ -320,21 +325,14 @@ public class TransactionSizeAndHashTest {
         .value(Wei.of(2000L))
         .payload(Bytes.fromHexString("0x12345678"))
         .maxFeePerBlobGas(Wei.of(250L))
-        .versionedHashes(
-            List.of(
-                new VersionedHash(
-                    Bytes32.fromHexString(
-                        "0x0122334455667788991011121314151617181920212223242526272829303101"))))
+        .versionedHashes(List.of(versionedHash))
         .blobsWithCommitments(
             new BlobsWithCommitments(
                 BlobType.KZG_PROOF,
-                List.of(new KZGCommitment(Bytes48.fromHexStringLenient("0x0987"))),
+                List.of(commitment),
                 List.of(new Blob(Bytes.fromHexString("0x0987"))),
                 List.of(new KZGProof(Bytes48.fromHexStringLenient("0x1234"))),
-                List.of(
-                    new VersionedHash(
-                        Bytes32.fromHexStringLenient(
-                            "0x0122334455667788991011121314151617181920212223242526272829303101")))))
+                List.of(versionedHash)))
         .signature(FAKE_SIGNATURE)
         .build();
   }