+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/transactions/NewPooledTransactionHashesMessageProcessor.java
@@ -93,6 +93,7 @@ public class NewPooledTransactionHashesMessageProcessor {
     try {
       final List<TransactionAnnouncement> incomingAnnouncements =
           transactionsMessage.pendingTransactionAnnouncements();
+
       final var freshAnnouncements =
           transactionTracker.receivedAnnouncements(peer, incomingAnnouncements);